Separated task into Task and BoardTask
